Fixes, new command and build.cmd
- fixed a bug in ScriptPack::Terminate where the logic wouldn't null out
the ClrMdPack instance if it wasn't attached
- changed DumpFinalizerQueueHeapStatsByType(minSize) from int to long
- added DumpStack command
- changed ConsoleOutputWriter::WriteLineSeparator to write only half the
console width for aesthetics
- added a build.cmd file to support easy building without having to fire
up Visual Studio

+++ src/ScriptCs.ClrMD/ClrMdPack.Commands.Heap.cs
@@ -40,7 +40,7 @@ namespace HackedBrain.ScriptCs.ClrMd
 			this.DumpFinalizerQueueHeapStatsByType(0, 0, null);
 		}
 
-		public void DumpFinalizerQueueHeapStatsByType(int minSize)
+		public void DumpFinalizerQueueHeapStatsByType(long minSize)
 		{
 			this.DumpFinalizerQueueHeapStatsByType(minSize, 0, null);
 		}
@@ -50,7 +50,7 @@ namespace HackedBrain.ScriptCs.ClrMd
 			this.DumpFinalizerQueueHeapStatsByType(0, 0, typeName);
 		}
 
-		public void DumpFinalizerQueueHeapStatsByType(int minSize, string typeName)
+		public void DumpFinalizerQueueHeapStatsByType(long minSize, string typeName)
 		{
 			this.DumpFinalizerQueueHeapStatsByType(minSize, 0, typeName);
 		}
